@wiki_mcp_fastmcp/pyproject.toml
[project]
name = "wiki-mcp-fastmcp"
version = "0.1.0"
description = "Confluence MCP Server implemented with FastMCP"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.14.3",
    "requests>=2.31.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
wiki-mcp-fastmcp = "wiki_mcp_fastmcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wiki_mcp_fastmcp"]
